% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{load_data}
\alias{load_data}
\title{n/a}
\usage{
load_data()
}
\value{
a data frame
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Sebastian Funk
}
