% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_trunc.r
\name{create_snapshots}
\alias{create_snapshots}
\title{n/a}
\usage{
create_snapshots(x)
}
\arguments{
\item{x}{data frame containing `report_date` and `date_onset` columns}
}
\value{
a list of data frames to feed into `estimate_truncation`
}
\description{
Create data snapshots from a reporting date
}
\details{
.. content for \details{} ..
}
\author{
Sebastian Funk
}
